Remove an import: javax.swing.event.ChangeListener. This is because javax.swing has a wildcard already.
Add set title to the JFrame.

Ln 155-156: Set X/Y positions so that the bars display properly on the JPanel.

Ln 158-159: Change calls to revalidate and repaint from JPanel barPanel to -> mainPanel. Reason: Instead of updating just the barPanel, it calls to update the entire mainPanel to reduce any visual oddities that were happening.

Remove unneeded lines in refreshBarWidth(). This is because the method was getting too complicated. It's now handled elsewhere.